+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/NoLineBreakAfterElseRule.kt
@@ -10,7 +10,7 @@ import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
 
-class NoLineBreakAfterElseRule : Rule("no-line-break-after-else") {
+public class NoLineBreakAfterElseRule : Rule("no-line-break-after-else") {
 
     override fun beforeVisitChildNodes(
         node: ASTNode,